Measure the operator boundary cones, and spend the slack they take back
Round 6 traced every fmax regression to one class: the critical path ends at
the first internal register of a pipelined float core, and the model let a mux
chain in front of it because every pipelined row inherited `in_delay_ns = 0.5`
from the archetype while the real entry cone is a whole mantissa path. Four
changes follow from measuring that number, plus the period dimension `-Oarea`
needed once the model was honest enough to rank slow clocks on.

Boundary cones (`IPRow.in_delay_ns` / new `out_delay_ns`). All 69 cores are
re-characterized on the shipped recipe, linked as a checkpoint into a
register-core-register wrapper so the wrapper's own registers cannot be
absorbed, and each of the three arcs timed apart. `out_delay_ns` did not exist
before and is the arc a chain leaving a core starts from, so it compares
directly against the register floor. Bed-wide this takes designs closing 300
MHz from 38 to 48.

Port select reservation. The select `bindMemoryPorts` grows in front of a
shared port bus had no term in the model the schedule is cut against.
`recordPortSelectArms` walks the module once, beside `recordArrayStorage`, and
records on every array access how many holders its (array, bank, direction)
bus may be coloured with; `portSelectDelay` prices them with the same
`muxCone` the realized STA uses and `accessCharacterization` charges it into
the access's setup delay, so no new constraint family is needed. The count is
the colouring's ceiling rather than accesses/ports, because the read merge
collapses up to 12 accesses onto one bus and the device's mux row saturates
past 8 arms, which makes the reservation a guarantee at almost no cost.

Balanced unit assignment. `headroomNs` reserves a select for `ceil(k / units)`
operations per instance, but `assignUnits` round-robined and could exceed it.
It now takes the least loaded free instance, and under a modulo schedule
places the fullest congruence classes first, since a class of c members needs
c distinct instances and leaving one for last forces it onto an instance the
balance has already filled. Spreading a class evenly also builds narrower
selects than packing it unevenly, so the mux LUT count falls.

Escaping literals. `materializeSequential` yielded an escaping `arith.constant`
as a region result, so the emitter latched it into a survivor register and a
shift by it built a barrel shifter where the schedule had priced wiring. A
declaration is not a value a region yields: literals are hoisted to func level
like an escaping alloc already was, and `spanEscapingValues` skips them so the
span objective and the reify select the same set.

`-Oarea` takes the period dimension. Since the cycles-order area fold landed,
both directions end in `min area s.t. span <= B` and differ only by the
heuristic's suboptimality, which the modulo rework drove to ~1.0022. So
`-Oarea` becomes a period policy like `-Owall`: `sweep_area` probes candidate
clocks, ranks them on a new module-wide modeled area (`ScheduleModel::
modeledArea`, the objective's own terms evaluated on a settled schedule), and
solves at the winner. Without a deadline the requested clock is the reference,
the rank is area times wall, and the winner is confirmed by a second solve at
the requested clock so the direction can never build more than it would have
without the sweep; `wall_ns` replaces that reference with a deadline and
refuses rather than overrun.

Cone delays are now stated on one grid (`kConeDelayQuantum`): a cone reaches a
schedule inside its consumer's operator type, whose name is what the problem
registers timing under, so a delay printed to two decimals has to be quantized
to two decimals or two cones printing alike would cost differently depending on
which was registered last. The address cone and the standalone apply cone move
onto it too.

Bed-wide against the pre-T1 state: designs missing a timing path 7 -> 1,
designs over the binding mux headroom 4 -> 0, LUT 0.9936, mux LUT 0.9890.
`tests/rtl` 500 passing and `benchmark.verify` 109 pass / 3 skip / 0 fail.

+++ tests/rtl/test_devices.py
@@ -10,7 +10,7 @@ import pytest
 
 from allo import kernel
 from allo.backend.rtl.core import RealizationWarning
-from allo.lang import f32, i32
+from allo.lang import f32, i32, i64
 from allo.lang.ip import operator_ip
 from allo.operators import arith as allo_arith
 from allo.operators import math as amath
@@ -178,8 +178,12 @@ def test_generate_builds_the_fabric_multiply_out_of_luts():
     # The multiply's fabric row spends no DSPs, so it rebuilds with the recipe's
     # DSP-free fragment, which repeats a key the shape sets and must land after
     # it; a `set_property -dict` list resolves the order.
+    #
+    # i64, because rank is latency before price: a fabric row only competes with
+    # the DSP row declared at its own depth, and at 32 bits the DSP row that
+    # holds this clock is a cycle shallower than the fabric one.
     @kernel
-    def mk(x: i32[8], y: i32[8], out: i32[8]):
+    def mk(x: i64[8], y: i64[8], out: i64[8]):
         for i in range(8):
             out[i] = x[i] * y[i]
